requirements.txt
altair==4.2.2
astor==0.8.1
attrs==23.1.0
backports.zoneinfo==0.2.1
base58==2.1.1
blinker==1.6.2
cachetools==5.3.1
certifi 
charset-normalizer==3.1.0
click==7.1.2
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
keras-contrib==2.0.8
MarkupSafe==2.1.3
numpy==1.21.6
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.19.1
pyarrow==12.0.1
pydeck==0.8.1b0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
smmap==5.0.0
streamlit==1.0.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wincertstore==0.2
zipp==3.15.0
